:zap: :: UseCaseImpl 정의

+++ Projects/Domain/ProductDomain/Sources/DTO/Request/CreateReviewRequestDTO.swift
@@ -11,6 +11,6 @@ import Foundation
 
 struct ReviewRequestDTO: Encodable {
     let content: String
-    let rating: Int
+    let rating: Float
     let images: [String]
 }
